test: 파일 구조 변경 및 테스트 대상 변경
