+++ server/jest.config.js
@@ -1,14 +1,14 @@
 module.exports = {
     moduleFileExtensions: ['js', 'json', 'ts'],
-    rootDir: 'src',
-    testRegex: '.*\\.spec\\.ts$',
+    rootDir: '.',
+    testRegex: 'test/.*\\.spec\\.ts$',
     transform: {
         '^.+\\.(t|j)s$': 'ts-jest',
     },
     collectCoverageFrom: [
-        '{auth,project,stt,user}/**/!(*.controller|*.module).ts',
-        '!auth/firebase-admin.ts',
+        'src/{auth,project,stt,user}/**/!(*.controller|*.module).ts',
+        '!src/auth/firebase-admin.ts',
     ],
-    coverageDirectory: '../coverage',
+    coverageDirectory: './coverage',
     testEnvironment: 'node',
 };
